requirements.txt
python-json-logger
opentelemetry-api